+++ internal/datasource_ct_config.go
@@ -1,4 +1,4 @@
-package ct
+package internal
 
 import (
 	"context"